Extract ConnectDomain component, reachable from publish dialog
The connect-a-domain flow was inline in the dashboard only. Extract it into a
reusable ConnectDomain component and add a "Connect a domain" button to the
publish success dialog for domainless sites, so users don't have to leave the
editor to make a site public.
