+++ src/lib/builder/views/editor/Toolbar.svelte
@@ -20,6 +20,7 @@
 	import PageTypeModal from '$lib/builder/views/modal/PageTypeModal/PageTypeModal.svelte'
 	import Collaboration from '$lib/builder/views/modal/Collaboration.svelte'
 	import Deploy from '$lib/components/Modals/Deploy/Deploy.svelte'
+	import ConnectDomain from '$lib/components/ConnectDomain.svelte'
 	import { usePublishSite } from '$lib/workers/Publish.svelte'
 	import { type Snippet } from 'svelte'
 	import { site_context } from '$lib/builder/stores/context'
@@ -115,6 +116,7 @@
 	let editing_collaborators = $state(false)
 	let publishing = $state(false)
 	let publish_stage = $state('INITIAL')
+	let connect_domain_open = $state(false)
 
 	// Close all dialogs on navigation
 	onNavigate(() => {
@@ -205,6 +207,11 @@
 			publish_fn={handle_publish}
 			loading={publish_in_progress}
 			site_host={site && is_host_assigned(site) ? site.host : ''}
+			onConnectDomain={() => {
+				publishing = false
+				publish_stage = 'INITIAL'
+				connect_domain_open = true
+			}}
 			onClose={() => {
 				publishing = false
 				publish_stage = 'INITIAL'
@@ -212,6 +219,8 @@
 		/>
 	</Dialog.Content>
 </Dialog.Root>
+
+<ConnectDomain {site} bind:open={connect_domain_open} />
 
 <nav aria-label="toolbar" id="primo-toolbar">
 	<div class="menu-container">
